revert original test cases and move new test cases to revisited package
