+++ modules/activiti-camel/src/test/java/org/activiti/camel/route/revisited/AsyncCamelRouteRevisted.java
@@ -11,21 +11,18 @@
  * limitations under the License.
  */
 
-package org.activiti.camel.route;
+package org.activiti.camel.route.revisited;
 
 import org.apache.camel.builder.RouteBuilder;
 
-public class AsyncCamelRoute extends RouteBuilder {
+public class AsyncCamelRouteRevisted extends RouteBuilder {
 
   @Override
   public void configure() throws Exception {
 
-    from("activiti:asyncCamelProcess:serviceTaskAsync1").setHeader("destination", constant("activiti:asyncCamelProcess:receive1")).to("seda:asyncQueue");
-    from("seda:asyncQueue").to("bean:sleepBean?method=sleep").to("seda:receiveQueue");
+    from("activiti:asyncCamelProcessRevisited:serviceTaskAsync1").to("bean:sleepBean?method=sleep").to("activiti:asyncCamelProcessRevisited:receive1");
     
-    from("activiti:asyncCamelProcess:serviceTaskAsync2").setHeader("destination", constant("activiti:asyncCamelProcess:receive2")).to("seda:asyncQueue2");
-    from("seda:asyncQueue2").to("bean:sleepBean?method=sleep").to("seda:receiveQueue");
-    
-    from("seda:receiveQueue").recipientList(header("destination"));
+    from("activiti:asyncCamelProcessRevisited:serviceTaskAsync2").to("bean:sleepBean?method=sleep").to("bean:sleepBean?method=sleep").to("activiti:asyncCamelProcessRevisited:receive2");    
+
   }
 }
