+++ modules/activiti-camel/src/test/resources/camel-activiti-context-revisited.xml
@@ -27,7 +27,7 @@
         <property name="dataSource" ref="dataSource"/>
         <property name="transactionManager" ref="transactionManager"/>
         <property name="databaseSchemaUpdate" value="true"/>
-        <property name="jobExecutorActivate" value="false"/>
+        <property name="jobExecutorActivate" value="true"/>
     </bean>
 
     <bean id="processEngine" class="org.activiti.spring.ProcessEngineFactoryBean">
