+++ modules/activiti-camel/src/test/resources/process/revisited/async-revisited.bpmn20.xml
@@ -8,18 +8,18 @@
              http://www.omg.org/spec/BPMN/20100524/MODEL http://www.omg.org/spec/BPMN/2.0/20100501/BPMN20.xsd">
 
 
-    <process id="asyncCamelProcess">
+    <process id="asyncCamelProcessRevisited">
 
         <startEvent id="start"/>
         <sequenceFlow sourceRef="start" targetRef="parallelSplit"/>
         <parallelGateway id="parallelSplit"/>
         <sequenceFlow sourceRef="parallelSplit" targetRef="serviceTaskAsync1"/>
         <sequenceFlow sourceRef="parallelSplit" targetRef="serviceTaskAsync2"/>
-        <serviceTask id="serviceTaskAsync1" activiti:type="camel"/>
+        <serviceTask id="serviceTaskAsync1" activiti:type="camel" activiti:async="true"/>
         <sequenceFlow sourceRef="serviceTaskAsync1" targetRef="receive1"/>
         <receiveTask id="receive1" name="Wait State" />
         <sequenceFlow sourceRef="receive1" targetRef="parallelJoin"/>
-        <serviceTask id="serviceTaskAsync2" activiti:type="camel"/>
+        <serviceTask id="serviceTaskAsync2" activiti:type="camel" activiti:async="true"/>
         <sequenceFlow sourceRef="serviceTaskAsync2" targetRef="receive2"/>
         <receiveTask id="receive2" name="Wait State" />
         <sequenceFlow sourceRef="receive2" targetRef="parallelJoin"/>
