+++ modules/activiti-engine/src/test/resources/activiti.cfg.xml
@@ -11,7 +11,6 @@
     <property name="jdbcDriver" value="org.h2.Driver" />
     <property name="jdbcUsername" value="sa" />
     <property name="jdbcPassword" value="" />
-    
 
     <!-- Database configurations -->
     <property name="databaseSchemaUpdate" value="drop-create" />
